# Optimize set! DSL (Round 2)