--- a/lib/jbuilder.rb
+++ b/lib/jbuilder.rb
@@ -239,8 +239,8 @@ class Jbuilder
 
   alias_method :method_missing, :set!
 
-  def _set(key, value = BLANK, attributes = nil)
-    result = if ::Kernel.block_given?
+  def _set(key, value = BLANK, attributes = nil, &block)
+    result = if block
       if _blank?(value)
         # json.comments { ... }
         # { "comments": ... }